Articles and Books Explaining the 1970s US Stagflation.

A dataset containing the books and academic articles endeavouring to explain what happened in the US economy in the 1970s, as well as all the articles and books cited at least twice by the first set of articles and books (on the stagflation).

Usage

Nodes_stagflation

Format

A data frame with 558 rows and 7 variables:

ItemID_Ref

Identifier of the document

Author

Author of the document

Author_date

Use this as a label for nodes

Year

Year of publication of the document

Title

Title of the document

Journal

Journal of publication of the document (if an article)

Type

If "Stagflation", the document is listed as an explanation of the US stagflation. If "Non-Stagflation", the document is cited by a document explaining the stagflation

Source

Goutsmedt A. (2020) “From Stagflation to the Great Inflation: Explaining the 1970s US Economic Situation”. Revue d’Economie Politique, Forthcoming 2021.