networkflow: Functions For A Workflow To Manipulate Networks

This package proposes a series of function to make it easier and quicker to work on networks. It mainly targets working on bibliometric networks (see the [biblionetwork](https://github.com/agoutsmedt/biblionetwork) package for creating such networks). This package heavily relies on [igraph](https://igraph.org/r/) and [tidygraph](https://tidygraph.data-imaginist.com/index.html), and aims at producing ready-made networks for projecting them using [ggraph](https://ggraph.data-imaginist.com/). This package does not invent nothing new, properly speaking, but it allows the users to follow more quickly and easily the main steps of network manipulation, from creating the graph to projecting it. It is inspired by what could be done with [GEPHI](https://gephi.org/): the package allows the use of the Leiden community detection algorithm, as well as of the Force Atlas 2 layout, both being unavailable in igraph (and so in tidygraph).

See also

Author

Maintainer: Aurélien Goutsmedt agoutsmedt@gmail.com (ORCID)

Authors:

Alexandre Truc alexandre.truc77@gmail.com (ORCID)