Find Similar Clusters across Multiple Temporal Networks

This function creates a new column "intertemporal_name" for each network from a list of temporal networks to identify similar clusters across time. The function gives the same name to two clusters from two succesive temporal networks if they match the conditions defined by the user: threshold_similarity, cluster_colum and similarity_type.

Usage

intertemporal_cluster_naming(
  list_graph = NA,
  cluster_column = NA,
  node_key = NA,
  threshold_similarity = 0.5001,
  similarity_type = c("complete, partial")
)

Arguments

list_graph

A list of networks. The list is expected to be ordered in a sequential order from the oldest to the most recent network.

cluster_column

The column with the identifier of the cluster.

node_key

The column with the unique identifier of each node.

threshold_similarity

The threshold_similarity variable defines how sensitive the function is to giving the same name to two clusters. A higher threshold will lead to more communities.

For example, if you have two temporal networks with two communities each. Communities A and B for the older network, and communities A' and B' for the more recent network. A threshold of 0.51 with a "complete" similarity_type means that community A' will be given the name A if 51% of the nodes from A' in the more recent network originate from A in the older network, and 51% of the node from A in the older network becomes in A' in the more recent network.

similarity_type

Choose a similarity type to compare the threshold to:

"complete" similarity compute the share of nodes going from a older community to a more recent community on all the nodes in both networks

"partial" similarity compute the share of nodes going from a older community to a more recent community only on nodes that exists in both networks

Complete similarity is particularly suited if the number of nodes in your networks is relatively stable over time as the threshold capture the share of all nodes moving between clusters. Partial similarity can be particularly useful when the number of nodes in your networks increases rapidly. The interpretation of the threshold is that it captures the share of nodes existing in both networks moving between clusters.

For example, with a complete similarity threshold of 0.51, if (1) all nodes from community A in network t-1 go into community A' in network t+1, and (2) all nodes in community A' present in network t-1 originate from community A, but (3) the number of nodes in A' is more than twice of A because of new nodes that did not exists in t-1, A' will never meet the threshold requirement to be named A despite a strong similarity between the two clusters. Conceptually, this might be a desired behavior of the function because one might considered that A' is too different from A to be considered the same cluster as its composition is changed from new nodes. In that case complete similarity is the right choice. However, if one consider that A and A' are very similar because all the nodes that exists in both networks are identified as part of the same community, then partial threshold similarity is more desirable.

Value

The function returns the same list of networks used as input in list_graph but with a new column intertemporal_name. The column is the result of the inter-temporal grouping of the original clusters of the cluster_column.

Examples

library(biblionetwork)
library(magrittr)
library(tidygraph)
#> 
#> Attaching package: ‘tidygraph’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter

nodes <- Nodes_stagflation %>%
dplyr::rename(ID_Art = ItemID_Ref) %>%
dplyr::filter(Type == "Stagflation")

references <- Ref_stagflation %>%
dplyr::rename(ID_Art = Citing_ItemID_Ref)

temporal_networks <- dynamic_network_cooccurrence(nodes = nodes,
directed_edges = references,
source_column = "ID_Art",
target_column = "ItemID_Ref",
time_variable = "Year",
cooccurrence_method = "coupling_similarity",
time_window = 15,
edges_threshold = 1,
compute_size = FALSE,
keep_singleton = FALSE,
overlapping_window = TRUE)
#> Creation of the network for the 1975-1989 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1976-1990 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1977-1991 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1978-1992 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1979-1993 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1980-1994 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1981-1995 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1982-1996 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1983-1997 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1984-1998 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1985-1999 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1986-2000 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1987-2001 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1988-2002 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1989-2003 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1990-2004 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1991-2005 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1992-2006 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1993-2007 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1994-2008 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1995-2009 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1996-2010 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1997-2011 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1998-2012 window.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1999-2013 window.
#> We remove the nodes that are alone with no edge. 
#> 

temporal_networks <- lapply(temporal_networks,
                                    function(tbl) tbl %N>%
                                                  mutate(clusters = tidygraph::group_louvain()))

intertemporal_cluster_naming(temporal_networks,
cluster_column = "clusters",
node_key = "ID_Art",
threshold_similarity = 0.51,
similarity_type = "partial")
#> Warning: `intertemporal_cluster_naming()` was deprecated in networkflow 0.1.0.
#> ℹ Please use `merge_dynamic_clusters()` instead.
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> Joining with `by = join_by(clusters)`
#> $`1975-1989`
#> # A tbl_graph: 71 nodes and 435 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 71 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 16182155   GORDON…  1975 GORDON-R-1… ALTE… "BROOK… Stag… 1975-1989   3       
#>  2 26283591   GORDON…  1975 GORDON-R-1… THE … "BROOK… Stag… 1975-1989   3       
#>  3 16182201   OKUN-A   1975 OKUN-A-197… INFL… "BROOK… Stag… 1975-1989   2       
#>  4 47749045   BRONFE…  1976 BRONFENBRE… ELEM… "ZEITS… Stag… 1975-1989   2       
#>  5 1111111141 KARNOS…  1976 KARNOSKY-D… THE … "REVIE… Stag… 1975-1989   1       
#>  6 1021902    FRIEDM…  1977 FRIEDMAN-M… NOBE… "THE J… Stag… 1975-1989   2       
#>  7 5200398    GOLDST…  1977 GOLDSTEIN-… DOWN… "STAFF… Stag… 1975-1989   3       
#>  8 31895842   GORDON…  1977 GORDON-R-1… CAN … "BROOK… Stag… 1975-1989   3       
#>  9 14371908   RASCHE…  1977 RASCHE-R-1… THE … ""      Stag… 1975-1989   4       
#> 10 6013999    SHERMA…  1977 SHERMAN-H-… MONO… "JOURN… Stag… 1975-1989   2       
#> # ℹ 61 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 435 × 5
#>    from    to   weight Source  Target    
#>   <int> <int>    <dbl> <chr>   <chr>     
#> 1     6    11 0.00158  1021902 1111111122
#> 2     6    45 0.000173 1021902 1111111128
#> 3     6    66 0.000430 1021902 1111111134
#> # ℹ 432 more rows
#> 
#> $`1976-1990`
#> # A tbl_graph: 66 nodes and 385 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 66 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 1111111141 KARNOS…  1976 KARNOSKY-D… THE … "REVIE… Stag… 1976-1990   3       
#>  2 1021902    FRIEDM…  1977 FRIEDMAN-M… NOBE… "THE J… Stag… 1976-1990   1       
#>  3 5200398    GOLDST…  1977 GOLDSTEIN-… DOWN… "STAFF… Stag… 1976-1990   1       
#>  4 31895842   GORDON…  1977 GORDON-R-1… CAN … "BROOK… Stag… 1976-1990   4       
#>  5 14371908   RASCHE…  1977 RASCHE-R-1… THE … ""      Stag… 1976-1990   2       
#>  6 1111111122 CARLSO…  1978 CARLSON-K-… INFL… "FEDER… Stag… 1976-1990   1       
#>  7 1111111131 ECKSTE…  1978 ECKSTEIN-O… THE … ""      Stag… 1976-1990   4       
#>  8 9756183    FAIR-R   1978 FAIR-R-1978 INFL… "AFTER… Stag… 1976-1990   1       
#>  9 1111111150 MORK-K   1978 MORK-K-197… THE … ""      Stag… 1976-1990   4       
#> 10 16182206   PERRY-G  1978 PERRY-G-19… SLOW… "BROOK… Stag… 1976-1990   4       
#> # ℹ 56 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 385 × 5
#>    from    to   weight Source  Target    
#>   <int> <int>    <dbl> <chr>   <chr>     
#> 1     2     6 0.00161  1021902 1111111122
#> 2     2    40 0.000174 1021902 1111111128
#> 3     2    61 0.000436 1021902 1111111134
#> # ℹ 382 more rows
#> 
#> $`1977-1991`
#> # A tbl_graph: 66 nodes and 384 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 66 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 1021902    FRIEDM…  1977 FRIEDMAN-M… NOBE… "THE J… Stag… 1977-1991   4       
#>  2 5200398    GOLDST…  1977 GOLDSTEIN-… DOWN… "STAFF… Stag… 1977-1991   2       
#>  3 31895842   GORDON…  1977 GORDON-R-1… CAN … "BROOK… Stag… 1977-1991   3       
#>  4 14371908   RASCHE…  1977 RASCHE-R-1… THE … ""      Stag… 1977-1991   1       
#>  5 1111111122 CARLSO…  1978 CARLSON-K-… INFL… "FEDER… Stag… 1977-1991   4       
#>  6 1111111131 ECKSTE…  1978 ECKSTEIN-O… THE … ""      Stag… 1977-1991   3       
#>  7 9756183    FAIR-R   1978 FAIR-R-1978 INFL… "AFTER… Stag… 1977-1991   4       
#>  8 1111111150 MORK-K   1978 MORK-K-197… THE … ""      Stag… 1977-1991   3       
#>  9 16182206   PERRY-G  1978 PERRY-G-19… SLOW… "BROOK… Stag… 1977-1991   3       
#> 10 2207592    PHELPS…  1978 PHELPS-E-1… COMM… "JOURN… Stag… 1977-1991   5       
#> # ℹ 56 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 384 × 5
#>    from    to   weight Source  Target    
#>   <int> <int>    <dbl> <chr>   <chr>     
#> 1     1     5 0.00161  1021902 1111111122
#> 2     1    39 0.000174 1021902 1111111128
#> 3     1    60 0.000436 1021902 1111111134
#> # ℹ 381 more rows
#> 
#> $`1978-1992`
#> # A tbl_graph: 61 nodes and 341 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 61 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 1111111122 CARLSO…  1978 CARLSON-K-… INFL… "FEDER… Stag… 1978-1992   1       
#>  2 1111111131 ECKSTE…  1978 ECKSTEIN-O… THE … ""      Stag… 1978-1992   4       
#>  3 9756183    FAIR-R   1978 FAIR-R-1978 INFL… "AFTER… Stag… 1978-1992   1       
#>  4 1111111150 MORK-K   1978 MORK-K-197… THE … ""      Stag… 1978-1992   4       
#>  5 16182206   PERRY-G  1978 PERRY-G-19… SLOW… "BROOK… Stag… 1978-1992   4       
#>  6 2207592    PHELPS…  1978 PHELPS-E-1… COMM… "JOURN… Stag… 1978-1992   1       
#>  7 16413171   STEIN-J  1978 STEIN-J-19… INFL… "JOURN… Stag… 1978-1992   1       
#>  8 1111111172 STEIN-J  1978 STEIN-J-19… INFL… "JOURN… Stag… 1978-1992   1       
#>  9 31302737   TATOM-J  1978 TATOM-J-19… DOES… "FEDER… Stag… 1978-1992   2       
#> 10 1111111180 WACHTE…  1978 WACHTER-M-… INST… ""      Stag… 1978-1992   1       
#> # ℹ 51 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 341 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    24    35 0.000497 10585974 1111111128
#> 2     2    24 0.000167 10585974 1111111131
#> 3    24    37 0.00188  10585974 1111111137
#> # ℹ 338 more rows
#> 
#> $`1979-1993`
#> # A tbl_graph: 51 nodes and 255 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 51 × 10 (active)
#>    ID_Art   Author     Year Author_date Title Journal Type  time_window clusters
#>    <chr>    <chr>     <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 5240516  BLINDER-A  1979 BLINDER-A-… ECON… ""      Stag… 1979-1993   3       
#>  2 6119013  BRUNO-M    1979 BRUNO-M-19… MACR… ""      Stag… 1979-1993   3       
#>  3 31502746 GRAMLICH…  1979 GRAMLICH-E… MACR… "BROOK… Stag… 1979-1993   4       
#>  4 18541214 NORSWORT…  1979 NORSWORTHY… THE … ""      Stag… 1979-1993   1       
#>  5 20623867 SACHS-J    1979 SACHS-J-19… WAGE… "BROOK… Stag… 1979-1993   2       
#>  6 33940681 TATOM-J    1979 TATOM-J-19… THE … ""      Stag… 1979-1993   1       
#>  7 41293703 WEINTRAU…  1979 WEINTRAUB-… COMM… "JOURN… Stag… 1979-1993   4       
#>  8 37095547 ATESOGLU…  1980 ATESOGLU-H… INFL… "JOURN… Stag… 1979-1993   4       
#>  9 1172921  BRANSON-W  1980 BRANSON-W-… INTE… ""      Stag… 1979-1993   2       
#> 10 14490177 BRUNNER-K  1980 BRUNNER-K-… STAG… "JOURN… Stag… 1979-1993   5       
#> # ℹ 41 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 255 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    13    23 0.000554 10585974 1111111128
#> 2    13    25 0.00203  10585974 1111111137
#> 3    13    19 0.00259  10585974 1111111160
#> # ℹ 252 more rows
#> 
#> $`1980-1994`
#> # A tbl_graph: 43 nodes and 171 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 43 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 37095547   ATESOG…  1980 ATESOGLU-H… INFL… "JOURN… Stag… 1980-1994   3       
#>  2 1172921    BRANSO…  1980 BRANSON-W-… INTE… ""      Stag… 1980-1994   4       
#>  3 14490177   BRUNNE…  1980 BRUNNER-K-… STAG… "JOURN… Stag… 1980-1994   4       
#>  4 14371905   BRUNO-M  1980 BRUNO-M-19… IMPO… "THE E… Stag… 1980-1994   1       
#>  5 10585974   FRYE-J   1980 FRYE-J-1980 THE … ""      Stag… 1980-1994   3       
#>  6 1111111146 LEIJON…  1980 LEIJONHUFV… THEO… "REVUE… Stag… 1980-1994   2       
#>  7 112214146  MORK-K   1980 MORK-K-198… ENER… ""      Stag… 1980-1994   4       
#>  8 25850485   NORDHA…  1980 NORDHAUS-W… OIL … "BROOK… Stag… 1980-1994   2       
#>  9 16008556   PERRY-G  1980 PERRY-G-19… INFL… "BROOK… Stag… 1980-1994   4       
#> 10 1111111159 PINDYC…  1980 PINDYCK-R-… ENER… "THE E… Stag… 1980-1994   2       
#> # ℹ 33 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 171 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1     5    15 0.000604 10585974 1111111128
#> 2     5    17 0.00222  10585974 1111111137
#> 3     5    11 0.00269  10585974 1111111160
#> # ℹ 168 more rows
#> 
#> $`1981-1995`
#> # A tbl_graph: 32 nodes and 106 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 32 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 25091445   BLINDE…  1981 BLINDER-A-… "THE… ""      Stag… 1981-1995   4       
#>  2 1111111128 DARBY-M  1981 DARBY-M-19… "THE… ""      Stag… 1981-1995   2       
#>  3 38951845   FRYE-J   1981 FRYE-J-1981 "GOV… "THE A… Stag… 1981-1995   4       
#>  4 1111111137 GORDON…  1981 GORDON-R-1… "INF… "NBER … Stag… 1981-1995   4       
#>  5 16001640   LUTZ-MA  1981 LUTZ-MA-19… "STA… ""      Stag… 1981-1995   1       
#>  6 5315078    RASCHE…  1981 RASCHE-R-1… "ENE… "CARNE… Stag… 1981-1995   2       
#>  7 46282251   ATESOG…  1982 ATESOGLU-H… "WAG… "JOURN… Stag… 1981-1995   4       
#>  8 42623802   BLAAS-W  1982 BLAAS-W-19… "INS… "JOURN… Stag… 1981-1995   4       
#>  9 31895841   BLINDE…  1982 BLINDER-A-… "THE… "INFLA… Stag… 1981-1995   4       
#> 10 1184127    DARBY-M  1982 DARBY-M-19… "THE… "THE A… Stag… 1981-1995   2       
#> # ℹ 22 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 106 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    29    32 0.000428 10729971 1111111176
#> 2    21    32 0.000428 10729971 16167977  
#> 3    30    32 0.00120  10729971 19627977  
#> # ℹ 103 more rows
#> 
#> $`1982-1996`
#> # A tbl_graph: 26 nodes and 78 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 26 × 10 (active)
#>    ID_Art   Author     Year Author_date Title Journal Type  time_window clusters
#>    <chr>    <chr>     <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 42623802 BLAAS-W    1982 BLAAS-W-19… INST… "JOURN… Stag… 1982-1996   1       
#>  2 31895841 BLINDER-A  1982 BLINDER-A-… THE … "INFLA… Stag… 1982-1996   2       
#>  3 1184127  DARBY-M    1982 DARBY-M-19… THE … "THE A… Stag… 1982-1996   1       
#>  4 14576298 GRUBB-D    1982 GRUBB-D-19… CAUS… "THE R… Stag… 1982-1996   1       
#>  5 33117553 KOURI-PJK  1982 KOURI-PJK-… MACR… "AMERI… Stag… 1982-1996   1       
#>  6 16995857 NORDHAUS…  1982 NORDHAUS-W… ECON… "EUROP… Stag… 1982-1996   3       
#>  7 18087990 OLSON-M    1982 OLSON-M-19… STAG… "AMERI… Stag… 1982-1996   3       
#>  8 7815498  DENISON-E  1983 DENISON-E-… THE … "THE E… Stag… 1982-1996   3       
#>  9 5982867  GYLFASON…  1983 GYLFASON-T… DOES… "CANAD… Stag… 1982-1996   1       
#> 10 4189938  LINDBECK…  1983 LINDBECK-A… THE … "THE E… Stag… 1982-1996   3       
#> # ℹ 16 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 78 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    22    25 0.000436 10729971 1111111176
#> 2    14    25 0.000436 10729971 16167977  
#> 3    23    25 0.00122  10729971 19627977  
#> # ℹ 75 more rows
#> 
#> $`1983-1997`
#> # A tbl_graph: 22 nodes and 69 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 22 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 7815498    DENISO…  1983 DENISON-E-… THE … "THE E… Stag… 1983-1997   2       
#>  2 5982867    GYLFAS…  1983 GYLFASON-T… DOES… "CANAD… Stag… 1983-1997   2       
#>  3 4189938    LINDBE…  1983 LINDBECK-A… THE … "THE E… Stag… 1983-1997   2       
#>  4 5160307    MCCALL…  1983 MCCALLUM-C… INFL… "THE E… Stag… 1983-1997   3       
#>  5 17500902   ROTEMB…  1983 ROTEMBERG-… SUPP… "JOURN… Stag… 1983-1997   3       
#>  6 2901822    BRUNO-M  1984 BRUNO-M-19… RAW … "THE Q… Stag… 1983-1997   2       
#>  7 16167977   GORDON…  1984 GORDON-R-1… SUPP… "THE A… Stag… 1983-1997   3       
#>  8 1111111134 FISCHE…  1985 FISCHER-S-… SUPP… "JOURN… Stag… 1983-1997   3       
#>  9 40564631   MARTIN…  1985 MARTIN-LW-… STAG… "AMERI… Stag… 1983-1997   3       
#> 10 1184130    GISSER…  1986 GISSER-M-1… CRUD… "JOURN… Stag… 1983-1997   2       
#> # ℹ 12 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 69 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    15    18 0.000353 10729971 1111111176
#> 2     7    18 0.000400 10729971 16167977  
#> 3    16    18 0.00123  10729971 19627977  
#> # ℹ 66 more rows
#> 
#> $`1984-1998`
#> # A tbl_graph: 19 nodes and 67 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 19 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 2901822    BRUNO-M  1984 BRUNO-M-19… RAW … "THE Q… Stag… 1984-1998   2       
#>  2 16167977   GORDON…  1984 GORDON-R-1… SUPP… "THE A… Stag… 1984-1998   3       
#>  3 1111111134 FISCHE…  1985 FISCHER-S-… SUPP… "JOURN… Stag… 1984-1998   2       
#>  4 40564631   MARTIN…  1985 MARTIN-LW-… STAG… "AMERI… Stag… 1984-1998   1       
#>  5 1184130    GISSER…  1986 GISSER-M-1… CRUD… "JOURN… Stag… 1984-1998   2       
#>  6 25481516   HAKES-…  1988 HAKES-DR-1… EVID… "REVIE… Stag… 1984-1998   3       
#>  7 6714008    HELLIW…  1988 HELLIWELL-… COMP… "JOURN… Stag… 1984-1998   2       
#>  8 1717556    OLSON-M  1988 OLSON-M-19… THE … "JOURN… Stag… 1984-1998   2       
#>  9 214927     BALL-L   1991 BALL-L-1991 THE … "JOURN… Stag… 1984-1998   3       
#> 10 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … "RESER… Stag… 1984-1998   1       
#> 11 19627977   PARKIN…  1993 PARKIN-M-1… INFL… "PRICE… Stag… 1984-1998   1       
#> 12 2207578    BALL-L   1995 BALL-L-199… RELA… "THE Q… Stag… 1984-1998   3       
#> 13 10729971   BALL-L   1995 BALL-L-199… TIME… "JOURN… Stag… 1984-1998   1       
#> 14 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1984-1998   2       
#> 15 43590636   BERNAN…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1984-1998   1       
#> 16 8456979    DELONG…  1997 DELONG-B-1… AMER… "REDUC… Stag… 1984-1998   3       
#> 17 27334211   TAYLOR…  1997 TAYLOR-J-1… AMER… "REDUC… Stag… 1984-1998   1       
#> 18 60300900   CHARI-V  1998 CHARI-V-19… EXPE… "JOURN… Stag… 1984-1998   1       
#> 19 17355017   MAYER-T  1998 MAYER-T-19… MONE… ""      Stag… 1984-1998   1       
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 67 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    10    13 0.000277 10729971 1111111176
#> 2     2    13 0.000353 10729971 16167977  
#> 3    13    19 0.00147  10729971 17355017  
#> # ℹ 64 more rows
#> 
#> $`1985-1999`
#> # A tbl_graph: 18 nodes and 65 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 18 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 1111111134 FISCHE…  1985 FISCHER-S-… SUPP… "JOURN… Stag… 1985-1999   2       
#>  2 40564631   MARTIN…  1985 MARTIN-LW-… STAG… "AMERI… Stag… 1985-1999   1       
#>  3 1184130    GISSER…  1986 GISSER-M-1… CRUD… "JOURN… Stag… 1985-1999   3       
#>  4 25481516   HAKES-…  1988 HAKES-DR-1… EVID… "REVIE… Stag… 1985-1999   2       
#>  5 6714008    HELLIW…  1988 HELLIWELL-… COMP… "JOURN… Stag… 1985-1999   2       
#>  6 214927     BALL-L   1991 BALL-L-1991 THE … "JOURN… Stag… 1985-1999   2       
#>  7 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … "RESER… Stag… 1985-1999   1       
#>  8 19627977   PARKIN…  1993 PARKIN-M-1… INFL… "PRICE… Stag… 1985-1999   3       
#>  9 2207578    BALL-L   1995 BALL-L-199… RELA… "THE Q… Stag… 1985-1999   3       
#> 10 10729971   BALL-L   1995 BALL-L-199… TIME… "JOURN… Stag… 1985-1999   1       
#> 11 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1985-1999   3       
#> 12 43590636   BERNAN…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1985-1999   3       
#> 13 8456979    DELONG…  1997 DELONG-B-1… AMER… "REDUC… Stag… 1985-1999   2       
#> 14 27334211   TAYLOR…  1997 TAYLOR-J-1… AMER… "REDUC… Stag… 1985-1999   1       
#> 15 60300900   CHARI-V  1998 CHARI-V-19… EXPE… "JOURN… Stag… 1985-1999   1       
#> 16 17355017   MAYER-T  1998 MAYER-T-19… MONE… ""      Stag… 1985-1999   1       
#> 17 73951174   IRELAN…  1999 IRELAND-P-… DOES… "JOURN… Stag… 1985-1999   1       
#> 18 24247525   TAYLOR…  1999 TAYLOR-J-1… A HI… "MONET… Stag… 1985-1999   1       
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 65 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1     7    10 0.000282 10729971 1111111176
#> 2    10    16 0.00149  10729971 17355017  
#> 3     8    10 0.00112  10729971 19627977  
#> # ℹ 62 more rows
#> 
#> $`1986-2000`
#> # A tbl_graph: 21 nodes and 107 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 21 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 1184130    GISSER…  1986 GISSER-M-1… CRUD… "JOURN… Stag… 1986-2000   3       
#>  2 25481516   HAKES-…  1988 HAKES-DR-1… EVID… "REVIE… Stag… 1986-2000   1       
#>  3 6714008    HELLIW…  1988 HELLIWELL-… COMP… "JOURN… Stag… 1986-2000   3       
#>  4 214927     BALL-L   1991 BALL-L-1991 THE … "JOURN… Stag… 1986-2000   1       
#>  5 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … "RESER… Stag… 1986-2000   2       
#>  6 19627977   PARKIN…  1993 PARKIN-M-1… INFL… "PRICE… Stag… 1986-2000   3       
#>  7 2207578    BALL-L   1995 BALL-L-199… RELA… "THE Q… Stag… 1986-2000   1       
#>  8 10729971   BALL-L   1995 BALL-L-199… TIME… "JOURN… Stag… 1986-2000   1       
#>  9 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1986-2000   3       
#> 10 43590636   BERNAN…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1986-2000   3       
#> # ℹ 11 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 107 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1     8    20 0.000132 10729971 1111111143
#> 2     5     8 0.000269 10729971 1111111176
#> 3     8    19 0.00102  10729971 15559333  
#> # ℹ 104 more rows
#> 
#> $`1987-2001`
#> # A tbl_graph: 24 nodes and 125 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 24 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 25481516   HAKES-…  1988 HAKES-DR-1… EVID… "REVIE… Stag… 1987-2001   3       
#>  2 6714008    HELLIW…  1988 HELLIWELL-… COMP… "JOURN… Stag… 1987-2001   2       
#>  3 1717556    OLSON-M  1988 OLSON-M-19… THE … "JOURN… Stag… 1987-2001   2       
#>  4 214927     BALL-L   1991 BALL-L-1991 THE … "JOURN… Stag… 1987-2001   3       
#>  5 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … "RESER… Stag… 1987-2001   3       
#>  6 19627977   PARKIN…  1993 PARKIN-M-1… INFL… "PRICE… Stag… 1987-2001   2       
#>  7 2207578    BALL-L   1995 BALL-L-199… RELA… "THE Q… Stag… 1987-2001   2       
#>  8 10729971   BALL-L   1995 BALL-L-199… TIME… "JOURN… Stag… 1987-2001   3       
#>  9 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1987-2001   2       
#> 10 43590636   BERNAN…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1987-2001   2       
#> # ℹ 14 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 125 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1     8    20 0.000133 10729971 1111111143
#> 2     5     8 0.000253 10729971 1111111176
#> 3     8    19 0.000992 10729971 15559333  
#> # ℹ 122 more rows
#> 
#> $`1988-2002`
#> # A tbl_graph: 32 nodes and 231 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 32 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 25481516   HAKES-…  1988 HAKES-DR-1… EVID… "REVIE… Stag… 1988-2002   2       
#>  2 6714008    HELLIW…  1988 HELLIWELL-… COMP… "JOURN… Stag… 1988-2002   2       
#>  3 1717556    OLSON-M  1988 OLSON-M-19… THE … "JOURN… Stag… 1988-2002   2       
#>  4 214927     BALL-L   1991 BALL-L-1991 THE … "JOURN… Stag… 1988-2002   2       
#>  5 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … "RESER… Stag… 1988-2002   1       
#>  6 19627977   PARKIN…  1993 PARKIN-M-1… INFL… "PRICE… Stag… 1988-2002   2       
#>  7 2207578    BALL-L   1995 BALL-L-199… RELA… "THE Q… Stag… 1988-2002   2       
#>  8 10729971   BALL-L   1995 BALL-L-199… TIME… "JOURN… Stag… 1988-2002   1       
#>  9 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1988-2002   2       
#> 10 43590636   BERNAN…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1988-2002   2       
#> # ℹ 22 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 231 × 5
#>    from    to   weight Source    Target    
#>   <int> <int>    <dbl> <chr>     <chr>     
#> 1    26    31 0.000840 101742296 105203321 
#> 2    20    31 0.000391 101742296 1111111143
#> 3    19    31 0.00251  101742296 15559333  
#> # ℹ 228 more rows
#> 
#> $`1989-2003`
#> # A tbl_graph: 34 nodes and 305 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 34 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 214927     BALL-L   1991 BALL-L-1991 THE … JOURNA… Stag… 1989-2003   2       
#>  2 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … RESERV… Stag… 1989-2003   2       
#>  3 19627977   PARKIN…  1993 PARKIN-M-1… INFL… PRICE … Stag… 1989-2003   2       
#>  4 2207578    BALL-L   1995 BALL-L-199… RELA… THE QU… Stag… 1989-2003   2       
#>  5 10729971   BALL-L   1995 BALL-L-199… TIME… JOURNA… Stag… 1989-2003   2       
#>  6 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… JOURNA… Stag… 1989-2003   3       
#>  7 43590636   BERNAN…  1997 BERNANKE-B… SYST… BROOKI… Stag… 1989-2003   3       
#>  8 8456979    DELONG…  1997 DELONG-B-1… AMER… REDUCI… Stag… 1989-2003   2       
#>  9 27334211   TAYLOR…  1997 TAYLOR-J-1… AMER… REDUCI… Stag… 1989-2003   2       
#> 10 60300900   CHARI-V  1998 CHARI-V-19… EXPE… JOURNA… Stag… 1989-2003   2       
#> # ℹ 24 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 305 × 5
#>    from    to   weight Source    Target    
#>   <int> <int>    <dbl> <chr>     <chr>     
#> 1    24    29 0.000796 101742296 105203321 
#> 2    18    29 0.000348 101742296 1111111143
#> 3    17    29 0.00233  101742296 15559333  
#> # ℹ 302 more rows
#> 
#> $`1990-2004`
#> # A tbl_graph: 39 nodes and 431 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 39 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 214927     BALL-L   1991 BALL-L-1991 THE … JOURNA… Stag… 1990-2004   2       
#>  2 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … RESERV… Stag… 1990-2004   2       
#>  3 19627977   PARKIN…  1993 PARKIN-M-1… INFL… PRICE … Stag… 1990-2004   2       
#>  4 2207578    BALL-L   1995 BALL-L-199… RELA… THE QU… Stag… 1990-2004   2       
#>  5 10729971   BALL-L   1995 BALL-L-199… TIME… JOURNA… Stag… 1990-2004   2       
#>  6 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… JOURNA… Stag… 1990-2004   3       
#>  7 43590636   BERNAN…  1997 BERNANKE-B… SYST… BROOKI… Stag… 1990-2004   3       
#>  8 8456979    DELONG…  1997 DELONG-B-1… AMER… REDUCI… Stag… 1990-2004   2       
#>  9 27334211   TAYLOR…  1997 TAYLOR-J-1… AMER… REDUCI… Stag… 1990-2004   2       
#> 10 60300900   CHARI-V  1998 CHARI-V-19… EXPE… JOURNA… Stag… 1990-2004   2       
#> # ℹ 29 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 431 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    29    38 0.00101  101742279 101742296
#> 2    37    38 0.000143 101742279 101777013
#> 3    38    39 0.000225 101742279 106566092
#> # ℹ 428 more rows
#> 
#> $`1991-2005`
#> # A tbl_graph: 47 nodes and 663 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 47 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 214927     BALL-L   1991 BALL-L-1991 THE … JOURNA… Stag… 1991-2005   2       
#>  2 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … RESERV… Stag… 1991-2005   2       
#>  3 19627977   PARKIN…  1993 PARKIN-M-1… INFL… PRICE … Stag… 1991-2005   2       
#>  4 2207578    BALL-L   1995 BALL-L-199… RELA… THE QU… Stag… 1991-2005   2       
#>  5 10729971   BALL-L   1995 BALL-L-199… TIME… JOURNA… Stag… 1991-2005   2       
#>  6 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… JOURNA… Stag… 1991-2005   3       
#>  7 43590636   BERNAN…  1997 BERNANKE-B… SYST… BROOKI… Stag… 1991-2005   3       
#>  8 8456979    DELONG…  1997 DELONG-B-1… AMER… REDUCI… Stag… 1991-2005   2       
#>  9 27334211   TAYLOR…  1997 TAYLOR-J-1… AMER… REDUCI… Stag… 1991-2005   2       
#> 10 60300900   CHARI-V  1998 CHARI-V-19… EXPE… JOURNA… Stag… 1991-2005   2       
#> # ℹ 37 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 663 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    29    38 0.000901 101742279 101742296
#> 2    37    38 0.000125 101742279 101777013
#> 3    38    41 0.000122 101742279 106371053
#> # ℹ 660 more rows
#> 
#> $`1992-2006`
#> # A tbl_graph: 50 nodes and 768 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 50 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 1111111176 TAYLOR…  1992 TAYLOR-J-1… THE … RESERV… Stag… 1992-2006   2       
#>  2 19627977   PARKIN…  1993 PARKIN-M-1… INFL… PRICE … Stag… 1992-2006   2       
#>  3 2207578    BALL-L   1995 BALL-L-199… RELA… THE QU… Stag… 1992-2006   2       
#>  4 10729971   BALL-L   1995 BALL-L-199… TIME… JOURNA… Stag… 1992-2006   2       
#>  5 24669774   ROTEMB…  1996 ROTEMBERG-… IMPE… JOURNA… Stag… 1992-2006   3       
#>  6 43590636   BERNAN…  1997 BERNANKE-B… SYST… BROOKI… Stag… 1992-2006   3       
#>  7 8456979    DELONG…  1997 DELONG-B-1… AMER… REDUCI… Stag… 1992-2006   2       
#>  8 27334211   TAYLOR…  1997 TAYLOR-J-1… AMER… REDUCI… Stag… 1992-2006   2       
#>  9 60300900   CHARI-V  1998 CHARI-V-19… EXPE… JOURNA… Stag… 1992-2006   2       
#> 10 94940407   HETZEL…  1998 HETZEL-R-1… ARTH… FEDERA… Stag… 1992-2006   1       
#> # ℹ 40 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 768 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    28    37 0.000882 101742279 101742296
#> 2    36    37 0.000123 101742279 101777013
#> 3    37    40 0.000115 101742279 106371053
#> # ℹ 765 more rows
#> 
#> $`1993-2007`
#> # A tbl_graph: 55 nodes and 967 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 55 × 10 (active)
#>    ID_Art   Author     Year Author_date Title Journal Type  time_window clusters
#>    <chr>    <chr>     <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 19627977 PARKIN-M   1993 PARKIN-M-1… INFL… "PRICE… Stag… 1993-2007   3       
#>  2 2207578  BALL-L     1995 BALL-L-199… RELA… "THE Q… Stag… 1993-2007   3       
#>  3 10729971 BALL-L     1995 BALL-L-199… TIME… "JOURN… Stag… 1993-2007   3       
#>  4 24669774 ROTEMBER…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1993-2007   2       
#>  5 43590636 BERNANKE…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1993-2007   2       
#>  6 8456979  DELONG-B   1997 DELONG-B-1… AMER… "REDUC… Stag… 1993-2007   3       
#>  7 27334211 TAYLOR-J   1997 TAYLOR-J-1… AMER… "REDUC… Stag… 1993-2007   3       
#>  8 60300900 CHARI-V    1998 CHARI-V-19… EXPE… "JOURN… Stag… 1993-2007   3       
#>  9 94940407 HETZEL-R   1998 HETZEL-R-1… ARTH… "FEDER… Stag… 1993-2007   1       
#> 10 17355017 MAYER-T    1998 MAYER-T-19… MONE… ""      Stag… 1993-2007   3       
#> # ℹ 45 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 967 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    27    36 0.000857 101742279 101742296
#> 2    35    36 0.000119 101742279 101777013
#> 3    36    39 0.000110 101742279 106371053
#> # ℹ 964 more rows
#> 
#> $`1994-2008`
#> # A tbl_graph: 55 nodes and 985 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 55 × 10 (active)
#>    ID_Art   Author     Year Author_date Title Journal Type  time_window clusters
#>    <chr>    <chr>     <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 2207578  BALL-L     1995 BALL-L-199… RELA… "THE Q… Stag… 1994-2008   3       
#>  2 10729971 BALL-L     1995 BALL-L-199… TIME… "JOURN… Stag… 1994-2008   3       
#>  3 24669774 ROTEMBER…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1994-2008   2       
#>  4 43590636 BERNANKE…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1994-2008   2       
#>  5 8456979  DELONG-B   1997 DELONG-B-1… AMER… "REDUC… Stag… 1994-2008   3       
#>  6 27334211 TAYLOR-J   1997 TAYLOR-J-1… AMER… "REDUC… Stag… 1994-2008   3       
#>  7 60300900 CHARI-V    1998 CHARI-V-19… EXPE… "JOURN… Stag… 1994-2008   3       
#>  8 94940407 HETZEL-R   1998 HETZEL-R-1… ARTH… "FEDER… Stag… 1994-2008   1       
#>  9 17355017 MAYER-T    1998 MAYER-T-19… MONE… ""      Stag… 1994-2008   3       
#> 10 73951174 IRELAND-P  1999 IRELAND-P-… DOES… "JOURN… Stag… 1994-2008   1       
#> # ℹ 45 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 985 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    26    35 0.000848 101742279 101742296
#> 2    34    35 0.000116 101742279 101777013
#> 3    35    38 0.000110 101742279 106371053
#> # ℹ 982 more rows
#> 
#> $`1995-2009`
#> # A tbl_graph: 60 nodes and 1151 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 60 × 10 (active)
#>    ID_Art   Author     Year Author_date Title Journal Type  time_window clusters
#>    <chr>    <chr>     <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 2207578  BALL-L     1995 BALL-L-199… RELA… "THE Q… Stag… 1995-2009   3       
#>  2 10729971 BALL-L     1995 BALL-L-199… TIME… "JOURN… Stag… 1995-2009   3       
#>  3 24669774 ROTEMBER…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1995-2009   2       
#>  4 43590636 BERNANKE…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1995-2009   2       
#>  5 8456979  DELONG-B   1997 DELONG-B-1… AMER… "REDUC… Stag… 1995-2009   3       
#>  6 27334211 TAYLOR-J   1997 TAYLOR-J-1… AMER… "REDUC… Stag… 1995-2009   3       
#>  7 60300900 CHARI-V    1998 CHARI-V-19… EXPE… "JOURN… Stag… 1995-2009   3       
#>  8 94940407 HETZEL-R   1998 HETZEL-R-1… ARTH… "FEDER… Stag… 1995-2009   1       
#>  9 17355017 MAYER-T    1998 MAYER-T-19… MONE… ""      Stag… 1995-2009   3       
#> 10 73951174 IRELAND-P  1999 IRELAND-P-… DOES… "JOURN… Stag… 1995-2009   1       
#> # ℹ 50 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 1,151 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    26    35 0.000838 101742279 101742296
#> 2    34    35 0.000115 101742279 101777013
#> 3    35    38 0.000108 101742279 106371053
#> # ℹ 1,148 more rows
#> 
#> $`1996-2010`
#> # A tbl_graph: 60 nodes and 1163 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 60 × 10 (active)
#>    ID_Art   Author     Year Author_date Title Journal Type  time_window clusters
#>    <chr>    <chr>     <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 24669774 ROTEMBER…  1996 ROTEMBERG-… IMPE… "JOURN… Stag… 1996-2010   2       
#>  2 43590636 BERNANKE…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1996-2010   2       
#>  3 8456979  DELONG-B   1997 DELONG-B-1… AMER… "REDUC… Stag… 1996-2010   3       
#>  4 27334211 TAYLOR-J   1997 TAYLOR-J-1… AMER… "REDUC… Stag… 1996-2010   3       
#>  5 60300900 CHARI-V    1998 CHARI-V-19… EXPE… "JOURN… Stag… 1996-2010   3       
#>  6 94940407 HETZEL-R   1998 HETZEL-R-1… ARTH… "FEDER… Stag… 1996-2010   1       
#>  7 17355017 MAYER-T    1998 MAYER-T-19… MONE… ""      Stag… 1996-2010   3       
#>  8 73951174 IRELAND-P  1999 IRELAND-P-… DOES… "JOURN… Stag… 1996-2010   1       
#>  9 24247525 TAYLOR-J   1999 TAYLOR-J-1… A HI… "MONET… Stag… 1996-2010   2       
#> 10 95020640 CHRISTIA…  2000 CHRISTIANO… THE … ""      Stag… 1996-2010   3       
#> # ℹ 50 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 1,163 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    24    33 0.000838 101742279 101742296
#> 2    32    33 0.000115 101742279 101777013
#> 3    33    36 0.000108 101742279 106371053
#> # ℹ 1,160 more rows
#> 
#> $`1997-2011`
#> # A tbl_graph: 60 nodes and 1183 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 60 × 10 (active)
#>    ID_Art   Author     Year Author_date Title Journal Type  time_window clusters
#>    <chr>    <chr>     <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 43590636 BERNANKE…  1997 BERNANKE-B… SYST… "BROOK… Stag… 1997-2011   2       
#>  2 8456979  DELONG-B   1997 DELONG-B-1… AMER… "REDUC… Stag… 1997-2011   3       
#>  3 27334211 TAYLOR-J   1997 TAYLOR-J-1… AMER… "REDUC… Stag… 1997-2011   3       
#>  4 60300900 CHARI-V    1998 CHARI-V-19… EXPE… "JOURN… Stag… 1997-2011   3       
#>  5 94940407 HETZEL-R   1998 HETZEL-R-1… ARTH… "FEDER… Stag… 1997-2011   1       
#>  6 17355017 MAYER-T    1998 MAYER-T-19… MONE… ""      Stag… 1997-2011   3       
#>  7 73951174 IRELAND-P  1999 IRELAND-P-… DOES… "JOURN… Stag… 1997-2011   3       
#>  8 24247525 TAYLOR-J   1999 TAYLOR-J-1… A HI… "MONET… Stag… 1997-2011   2       
#>  9 95020640 CHRISTIA…  2000 CHRISTIANO… THE … ""      Stag… 1997-2011   1       
#> 10 76064614 CLARIDA-R  2000 CLARIDA-R-… MONE… "THE Q… Stag… 1997-2011   2       
#> # ℹ 50 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 1,183 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    23    32 0.000836 101742279 101742296
#> 2    31    32 0.000114 101742279 101777013
#> 3    32    35 0.000107 101742279 106371053
#> # ℹ 1,180 more rows
#> 
#> $`1998-2012`
#> # A tbl_graph: 59 nodes and 1199 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 59 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 60300900   CHARI-V  1998 CHARI-V-19… EXPE… "JOURN… Stag… 1998-2012   1       
#>  2 94940407   HETZEL…  1998 HETZEL-R-1… ARTH… "FEDER… Stag… 1998-2012   2       
#>  3 17355017   MAYER-T  1998 MAYER-T-19… MONE… ""      Stag… 1998-2012   1       
#>  4 73951174   IRELAN…  1999 IRELAND-P-… DOES… "JOURN… Stag… 1998-2012   1       
#>  5 24247525   TAYLOR…  1999 TAYLOR-J-1… A HI… "MONET… Stag… 1998-2012   3       
#>  6 95020640   CHRIST…  2000 CHRISTIANO… THE … ""      Stag… 1998-2012   2       
#>  7 76064614   CLARID…  2000 CLARIDA-R-… MONE… "THE Q… Stag… 1998-2012   3       
#>  8 15559333   LANSIN…  2000 LANSING-K-… EXPL… "RESEA… Stag… 1998-2012   1       
#>  9 1111111143 LANSIN…  2000 LANSING-K-… LEAR… ""      Stag… 1998-2012   1       
#> 10 19381502   ORPHAN…  2000 ORPHANIDES… ACTI… ""      Stag… 1998-2012   1       
#> # ℹ 49 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 1,199 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    20    29 0.000835 101742279 101742296
#> 2    28    29 0.000112 101742279 101777013
#> 3    29    32 0.000105 101742279 106371053
#> # ℹ 1,196 more rows
#> 
#> $`1999-2013`
#> # A tbl_graph: 64 nodes and 1515 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 64 × 10 (active)
#>    ID_Art     Author   Year Author_date Title Journal Type  time_window clusters
#>    <chr>      <chr>   <int> <chr>       <chr> <chr>   <chr> <chr>       <chr>   
#>  1 73951174   IRELAN…  1999 IRELAND-P-… DOES… "JOURN… Stag… 1999-2013   1       
#>  2 24247525   TAYLOR…  1999 TAYLOR-J-1… A HI… "MONET… Stag… 1999-2013   2       
#>  3 95020640   CHRIST…  2000 CHRISTIANO… THE … ""      Stag… 1999-2013   3       
#>  4 76064614   CLARID…  2000 CLARIDA-R-… MONE… "THE Q… Stag… 1999-2013   2       
#>  5 15559333   LANSIN…  2000 LANSING-K-… EXPL… "RESEA… Stag… 1999-2013   1       
#>  6 1111111143 LANSIN…  2000 LANSING-K-… LEAR… ""      Stag… 1999-2013   1       
#>  7 19381502   ORPHAN…  2000 ORPHANIDES… ACTI… ""      Stag… 1999-2013   1       
#>  8 95575922   COGLEY…  2001 COGLEY-T-2… EVOL… "NBER … Stag… 1999-2013   1       
#>  9 82181005   DAVIS-S  2001 DAVIS-S-20… SECT… "JOURN… Stag… 1999-2013   2       
#> 10 81273149   NITZAN…  2001 NITZAN-J-2… REGI… "REVIE… Stag… 1999-2013   2       
#> # ℹ 54 more rows
#> # ℹ 1 more variable: intertemporal_name <chr>
#> #
#> # Edge Data: 1,515 × 5
#>    from    to    weight Source    Target   
#>   <int> <int>     <dbl> <chr>     <chr>    
#> 1    17    26 0.000785  101742279 101742296
#> 2    25    26 0.000105  101742279 101777013
#> 3    26    29 0.0000994 101742279 106371053
#> # ℹ 1,512 more rows
#>