Find Similar Clusters across Multiple Temporal Networks

This function creates a new column "intertemporal_name" for each network from a list of temporal networks to identify similar clusters across time. The function gives the same name to two clusters from two succesive temporal networks if they match the conditions defined by the user: threshold_similarity, cluster_colum and similarity_type.

Usage

intertemporal_cluster_naming(
  list_graph = NA,
  cluster_column = NA,
  node_key = NA,
  threshold_similarity = 0.5001,
  similarity_type = c("complete, partial")
)

Arguments

list_graph

A list of networks. The list is expected to be ordered in a sequential order from the oldest to the most recent network.

cluster_column

The column with the identifier of the cluster.

node_key

The column with the unique identifier of each node.

threshold_similarity

The threshold_similarity variable defines how sensitive the function is to giving the same name to two clusters. A higher threshold will lead to more communities.

For example, if you have two temporal networks with two communities each. Communities A and B for the older network, and communities A' and B' for the more recent network. A threshold of 0.51 with a "complete" similarity_type means that community A' will be given the name A if 51% of the nodes from A' in the more recent network originate from A in the older network, and 51% of the node from A in the older network becomes in A' in the more recent network.

similarity_type

Choose a similarity type to compare the threshold to:

"complete" similarity compute the share of nodes going from a older community to a more recent community on all the nodes in both networks

"partial" similarity compute the share of nodes going from a older community to a more recent community only on nodes that exists in both networks

Complete similarity is particularly suited if the number of nodes in your networks is relatively stable over time as the threshold capture the share of all nodes moving between clusters. Partial similarity can be particularly useful when the number of nodes in your networks increases rapidly. The interpretation of the threshold is that it captures the share of nodes existing in both networks moving between clusters.

For example, with a complete similarity threshold of 0.51, if (1) all nodes from community A in network t-1 go into community A' in network t+1, and (2) all nodes in community A' present in network t-1 originate from community A, but (3) the number of nodes in A' is more than twice of A because of new nodes that did not exists in t-1, A' will never meet the threshold requirement to be named A despite a strong similarity between the two clusters. Conceptually, this might be a desired behavior of the function because one might considered that A' is too different from A to be considered the same cluster as its composition is changed from new nodes. In that case complete similarity is the right choice. However, if one consider that A and A' are very similar because all the nodes that exists in both networks are identified as part of the same community, then partial threshold similarity is more desirable.

Value

The function returns the same list of networks used as input in list_graph but with a new column intertemporal_name. The column is the result of the inter-temporal grouping of the original clusters of the cluster_column.

Examples

library(biblionetwork)
library(magrittr)
library(tidygraph)
#> 
#> Attaching package: ‘tidygraph’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter

nodes <- Nodes_stagflation %>%
dplyr::rename(ID_Art = ItemID_Ref) %>%
dplyr::filter(Type == "Stagflation")

references <- Ref_stagflation %>%
dplyr::rename(ID_Art = Citing_ItemID_Ref)

temporal_networks <- dynamic_network_cooccurrence(nodes = nodes,
directed_edges = references,
source_column = "ID_Art",
target_column = "ItemID_Ref",
time_variable = "Year",
cooccurrence_method = "coupling_similarity",
time_window = 15,
edges_threshold = 1,
compute_size = FALSE,
keep_singleton = FALSE,
overlapping_window = TRUE)
#> Creation of the network for the 1975-1989 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1976-1990 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1977-1991 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1978-1992 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1979-1993 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1980-1994 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1981-1995 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1982-1996 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1983-1997 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1984-1998 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1985-1999 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1986-2000 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1987-2001 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1988-2002 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1989-2003 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1990-2004 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1991-2005 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1992-2006 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1993-2007 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1994-2008 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1995-2009 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1996-2010 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1997-2011 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1998-2012 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 
#> Creation of the network for the 1999-2013 window.
#> The method use for co-occurence is the coupling_similarity method. The edge threshold is:1.
#> We remove the nodes that are alone with no edge. 
#> 

temporal_networks <- lapply(temporal_networks,
                                    function(tbl) tbl %N>%
                                                  mutate(clusters = tidygraph::group_louvain()))

intertemporal_cluster_naming(temporal_networks,
cluster_column = "clusters",
node_key = "ID_Art",
threshold_similarity = 0.51,
similarity_type = "partial")
#> Warning: `intertemporal_cluster_naming()` was deprecated in networkflow 0.1.0.
#> ℹ Please use `merge_dynamic_clusters()` instead.
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> Joining, by = "clusters"
#> $`1975-1989`
#> # A tbl_graph: 71 nodes and 435 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 71 × 10 (active)
#>   ID_Art     Author   Year Author… Title  Journal  Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>   <chr>  <chr>    <chr> <chr>   <chr>   <chr>  
#> 1 16182155   GORDON…  1975 GORDON… ALTER… "BROOKI… Stag… 1975-1… 3       cl_3   
#> 2 26283591   GORDON…  1975 GORDON… THE I… "BROOKI… Stag… 1975-1… 3       cl_3   
#> 3 16182201   OKUN-A   1975 OKUN-A… INFLA… "BROOKI… Stag… 1975-1… 4       cl_4   
#> 4 47749045   BRONFE…  1976 BRONFE… ELEME… "ZEITSC… Stag… 1975-1… 4       cl_4   
#> 5 1111111141 KARNOS…  1976 KARNOS… THE L… "REVIEW" Stag… 1975-1… 1       cl_1   
#> 6 1021902    FRIEDM…  1977 FRIEDM… NOBEL… "THE JO… Stag… 1975-1… 4       cl_4   
#> # … with 65 more rows
#> #
#> # Edge Data: 435 × 5
#>    from    to   weight Source  Target    
#>   <int> <int>    <dbl> <chr>   <chr>     
#> 1     6    11 0.00158  1021902 1111111122
#> 2     6    45 0.000173 1021902 1111111128
#> 3     6    66 0.000430 1021902 1111111134
#> # … with 432 more rows
#> 
#> $`1976-1990`
#> # A tbl_graph: 66 nodes and 385 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 66 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 1111111141 KARNOS…  1976 KARNOSK… THE L… "REVIE… Stag… 1976-1… 4       cl_1   
#> 2 1021902    FRIEDM…  1977 FRIEDMA… NOBEL… "THE J… Stag… 1976-1… 5       cl_4   
#> 3 5200398    GOLDST…  1977 GOLDSTE… DOWNW… "STAFF… Stag… 1976-1… 2       cl_5   
#> 4 31895842   GORDON…  1977 GORDON-… CAN T… "BROOK… Stag… 1976-1… 3       cl_3   
#> 5 14371908   RASCHE…  1977 RASCHE-… THE E… ""      Stag… 1976-1… 1       cl_2   
#> 6 1111111122 CARLSO…  1978 CARLSON… INFLA… "FEDER… Stag… 1976-1… 5       cl_4   
#> # … with 60 more rows
#> #
#> # Edge Data: 385 × 5
#>    from    to   weight Source  Target    
#>   <int> <int>    <dbl> <chr>   <chr>     
#> 1     2     6 0.00161  1021902 1111111122
#> 2     2    40 0.000174 1021902 1111111128
#> 3     2    61 0.000436 1021902 1111111134
#> # … with 382 more rows
#> 
#> $`1977-1991`
#> # A tbl_graph: 66 nodes and 384 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 66 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 1021902    FRIEDM…  1977 FRIEDMA… NOBEL… "THE J… Stag… 1977-1… 1       cl_5   
#> 2 5200398    GOLDST…  1977 GOLDSTE… DOWNW… "STAFF… Stag… 1977-1… 1       cl_5   
#> 3 31895842   GORDON…  1977 GORDON-… CAN T… "BROOK… Stag… 1977-1… 4       cl_3   
#> 4 14371908   RASCHE…  1977 RASCHE-… THE E… ""      Stag… 1977-1… 2       cl_2   
#> 5 1111111122 CARLSO…  1978 CARLSON… INFLA… "FEDER… Stag… 1977-1… 1       cl_5   
#> 6 1111111131 ECKSTE…  1978 ECKSTEI… THE G… ""      Stag… 1977-1… 4       cl_3   
#> # … with 60 more rows
#> #
#> # Edge Data: 384 × 5
#>    from    to   weight Source  Target    
#>   <int> <int>    <dbl> <chr>   <chr>     
#> 1     1     5 0.00161  1021902 1111111122
#> 2     1    39 0.000174 1021902 1111111128
#> 3     1    60 0.000436 1021902 1111111134
#> # … with 381 more rows
#> 
#> $`1978-1992`
#> # A tbl_graph: 61 nodes and 341 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 61 × 10 (active)
#>   ID_Art     Author   Year Author… Title  Journal  Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>   <chr>  <chr>    <chr> <chr>   <chr>   <chr>  
#> 1 1111111122 CARLSO…  1978 CARLSO… INFLA… "FEDERA… Stag… 1978-1… 1       cl_5   
#> 2 1111111131 ECKSTE…  1978 ECKSTE… THE G… ""       Stag… 1978-1… 3       cl_3   
#> 3 9756183    FAIR-R   1978 FAIR-R… INFLA… "AFTER … Stag… 1978-1… 1       cl_5   
#> 4 1111111150 MORK-K   1978 MORK-K… THE I… ""       Stag… 1978-1… 3       cl_3   
#> 5 16182206   PERRY-G  1978 PERRY-… SLOWI… "BROOKI… Stag… 1978-1… 3       cl_3   
#> 6 2207592    PHELPS…  1978 PHELPS… COMMO… "JOURNA… Stag… 1978-1… 1       cl_5   
#> # … with 55 more rows
#> #
#> # Edge Data: 341 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    24    35 0.000497 10585974 1111111128
#> 2     2    24 0.000167 10585974 1111111131
#> 3    24    37 0.00188  10585974 1111111137
#> # … with 338 more rows
#> 
#> $`1979-1993`
#> # A tbl_graph: 51 nodes and 255 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 51 × 10 (active)
#>   ID_Art   Author   Year Author_… Title   Journal  Type  time_w… cluste… intert…
#>   <chr>    <chr>   <int> <chr>    <chr>   <chr>    <chr> <chr>   <chr>   <chr>  
#> 1 5240516  BLINDE…  1979 BLINDER… ECONOM… ""       Stag… 1979-1… 3       cl_3   
#> 2 6119013  BRUNO-M  1979 BRUNO-M… MACRO-… ""       Stag… 1979-1… 5       cl_21  
#> 3 31502746 GRAMLI…  1979 GRAMLIC… MACRO … "BROOKI… Stag… 1979-1… 4       cl_18  
#> 4 18541214 NORSWO…  1979 NORSWOR… THE SL… ""       Stag… 1979-1… 2       cl_2   
#> 5 20623867 SACHS-J  1979 SACHS-J… WAGES,… "BROOKI… Stag… 1979-1… 1       cl_24  
#> 6 33940681 TATOM-J  1979 TATOM-J… THE PR… ""       Stag… 1979-1… 2       cl_2   
#> # … with 45 more rows
#> #
#> # Edge Data: 255 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    13    23 0.000554 10585974 1111111128
#> 2    13    25 0.00203  10585974 1111111137
#> 3    13    19 0.00259  10585974 1111111160
#> # … with 252 more rows
#> 
#> $`1980-1994`
#> # A tbl_graph: 43 nodes and 171 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 43 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 37095547   ATESOG…  1980 ATESOGL… INFLA… "JOURN… Stag… 1980-1… 3       cl_18  
#> 2 1172921    BRANSO…  1980 BRANSON… INTER… ""      Stag… 1980-1… 3       cl_18  
#> 3 14490177   BRUNNE…  1980 BRUNNER… STAGF… "JOURN… Stag… 1980-1… 1       cl_24  
#> 4 14371905   BRUNO-M  1980 BRUNO-M… IMPOR… "THE E… Stag… 1980-1… 2       cl_2   
#> 5 10585974   FRYE-J   1980 FRYE-J-… THE V… ""      Stag… 1980-1… 4       cl_3   
#> 6 1111111146 LEIJON…  1980 LEIJONH… THEOR… "REVUE… Stag… 1980-1… 1       cl_24  
#> # … with 37 more rows
#> #
#> # Edge Data: 171 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1     5    15 0.000604 10585974 1111111128
#> 2     5    17 0.00222  10585974 1111111137
#> 3     5    11 0.00269  10585974 1111111160
#> # … with 168 more rows
#> 
#> $`1981-1995`
#> # A tbl_graph: 32 nodes and 106 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 32 × 10 (active)
#>   ID_Art     Author     Year Author… Title Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>     <int> <chr>   <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 25091445   BLINDER-A  1981 BLINDE… "THE… ""      Stag… 1981-1… 4       cl_3   
#> 2 1111111128 DARBY-M    1981 DARBY-… "THE… ""      Stag… 1981-1… 2       cl_30  
#> 3 38951845   FRYE-J     1981 FRYE-J… "GOV… "THE A… Stag… 1981-1… 4       cl_3   
#> 4 1111111137 GORDON-R   1981 GORDON… "INF… "NBER … Stag… 1981-1… 4       cl_3   
#> 5 16001640   LUTZ-MA    1981 LUTZ-M… "STA… ""      Stag… 1981-1… 3       cl_31  
#> 6 5315078    RASCHE-R   1981 RASCHE… "ENE… "CARNE… Stag… 1981-1… 1       cl_2   
#> # … with 26 more rows
#> #
#> # Edge Data: 106 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    29    32 0.000428 10729971 1111111176
#> 2    21    32 0.000428 10729971 16167977  
#> 3    30    32 0.00120  10729971 19627977  
#> # … with 103 more rows
#> 
#> $`1982-1996`
#> # A tbl_graph: 26 nodes and 78 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 26 × 10 (active)
#>   ID_Art   Author      Year Author_… Title Journal Type  time_w… cluste… intert…
#>   <chr>    <chr>      <int> <chr>    <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 42623802 BLAAS-W     1982 BLAAS-W… INST… "JOURN… Stag… 1982-1… 1       cl_30  
#> 2 31895841 BLINDER-A   1982 BLINDER… THE … "INFLA… Stag… 1982-1… 2       cl_31  
#> 3 1184127  DARBY-M     1982 DARBY-M… THE … "THE A… Stag… 1982-1… 1       cl_30  
#> 4 14576298 GRUBB-D     1982 GRUBB-D… CAUS… "THE R… Stag… 1982-1… 1       cl_30  
#> 5 33117553 KOURI-PJK   1982 KOURI-P… MACR… "AMERI… Stag… 1982-1… 1       cl_30  
#> 6 16995857 NORDHAUS-W  1982 NORDHAU… ECON… "EUROP… Stag… 1982-1… 3       cl_2   
#> # … with 20 more rows
#> #
#> # Edge Data: 78 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    22    25 0.000436 10729971 1111111176
#> 2    14    25 0.000436 10729971 16167977  
#> 3    23    25 0.00122  10729971 19627977  
#> # … with 75 more rows
#> 
#> $`1983-1997`
#> # A tbl_graph: 22 nodes and 69 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 22 × 10 (active)
#>   ID_Art   Author       Year Author… Title Journal Type  time_w… cluste… intert…
#>   <chr>    <chr>       <int> <chr>   <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 7815498  DENISON-E    1983 DENISO… THE … "THE E… Stag… 1983-1… 2       cl_2   
#> 2 5982867  GYLFASON-T   1983 GYLFAS… DOES… "CANAD… Stag… 1983-1… 2       cl_2   
#> 3 4189938  LINDBECK-A   1983 LINDBE… THE … "THE E… Stag… 1983-1… 2       cl_2   
#> 4 5160307  MCCALLUM-C   1983 MCCALL… INFL… "THE E… Stag… 1983-1… 3       cl_30  
#> 5 17500902 ROTEMBERG-J  1983 ROTEMB… SUPP… "JOURN… Stag… 1983-1… 3       cl_30  
#> 6 2901822  BRUNO-M      1984 BRUNO-… RAW … "THE Q… Stag… 1983-1… 2       cl_2   
#> # … with 16 more rows
#> #
#> # Edge Data: 69 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    15    18 0.000353 10729971 1111111176
#> 2     7    18 0.000400 10729971 16167977  
#> 3    16    18 0.00123  10729971 19627977  
#> # … with 66 more rows
#> 
#> $`1984-1998`
#> # A tbl_graph: 19 nodes and 67 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 19 × 10 (active)
#>   ID_Art     Author     Year Author… Title Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>     <int> <chr>   <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 2901822    BRUNO-M    1984 BRUNO-… RAW … "THE Q… Stag… 1984-1… 2       cl_30  
#> 2 16167977   GORDON-R   1984 GORDON… SUPP… "THE A… Stag… 1984-1… 2       cl_30  
#> 3 1111111134 FISCHER-S  1985 FISCHE… SUPP… "JOURN… Stag… 1984-1… 2       cl_30  
#> 4 40564631   MARTIN-LW  1985 MARTIN… STAG… "AMERI… Stag… 1984-1… 1       cl_31  
#> 5 1184130    GISSER-M   1986 GISSER… CRUD… "JOURN… Stag… 1984-1… 3       cl_41  
#> 6 25481516   HAKES-DR   1988 HAKES-… EVID… "REVIE… Stag… 1984-1… 2       cl_30  
#> # … with 13 more rows
#> #
#> # Edge Data: 67 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1    10    13 0.000277 10729971 1111111176
#> 2     2    13 0.000353 10729971 16167977  
#> 3    13    19 0.00147  10729971 17355017  
#> # … with 64 more rows
#> 
#> $`1985-1999`
#> # A tbl_graph: 18 nodes and 65 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 18 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 1111111134 FISCHE…  1985 FISCHER… SUPPL… "JOURN… Stag… 1985-1… 2       cl_41  
#> 2 40564631   MARTIN…  1985 MARTIN-… STAGF… "AMERI… Stag… 1985-1… 1       cl_31  
#> 3 1184130    GISSER…  1986 GISSER-… CRUDE… "JOURN… Stag… 1985-1… 2       cl_41  
#> 4 25481516   HAKES-…  1988 HAKES-D… EVIDE… "REVIE… Stag… 1985-1… 3       cl_44  
#> 5 6714008    HELLIW…  1988 HELLIWE… COMPA… "JOURN… Stag… 1985-1… 2       cl_41  
#> 6 214927     BALL-L   1991 BALL-L-… THE G… "JOURN… Stag… 1985-1… 3       cl_44  
#> # … with 12 more rows
#> #
#> # Edge Data: 65 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1     7    10 0.000282 10729971 1111111176
#> 2    10    16 0.00149  10729971 17355017  
#> 3     8    10 0.00112  10729971 19627977  
#> # … with 62 more rows
#> 
#> $`1986-2000`
#> # A tbl_graph: 21 nodes and 107 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 21 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 1184130    GISSER…  1986 GISSER-… CRUDE… "JOURN… Stag… 1986-2… 2       cl_41  
#> 2 25481516   HAKES-…  1988 HAKES-D… EVIDE… "REVIE… Stag… 1986-2… 3       cl_44  
#> 3 6714008    HELLIW…  1988 HELLIWE… COMPA… "JOURN… Stag… 1986-2… 2       cl_41  
#> 4 214927     BALL-L   1991 BALL-L-… THE G… "JOURN… Stag… 1986-2… 3       cl_44  
#> 5 1111111176 TAYLOR…  1992 TAYLOR-… THE G… "RESER… Stag… 1986-2… 1       cl_31  
#> 6 19627977   PARKIN…  1993 PARKIN-… INFLA… "PRICE… Stag… 1986-2… 2       cl_41  
#> # … with 15 more rows
#> #
#> # Edge Data: 107 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1     8    20 0.000132 10729971 1111111143
#> 2     5     8 0.000269 10729971 1111111176
#> 3     8    19 0.00102  10729971 15559333  
#> # … with 104 more rows
#> 
#> $`1987-2001`
#> # A tbl_graph: 24 nodes and 125 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 24 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 25481516   HAKES-…  1988 HAKES-D… EVIDE… "REVIE… Stag… 1987-2… 1       cl_48  
#> 2 6714008    HELLIW…  1988 HELLIWE… COMPA… "JOURN… Stag… 1987-2… 3       cl_49  
#> 3 1717556    OLSON-M  1988 OLSON-M… THE P… "JOURN… Stag… 1987-2… 3       cl_49  
#> 4 214927     BALL-L   1991 BALL-L-… THE G… "JOURN… Stag… 1987-2… 1       cl_48  
#> 5 1111111176 TAYLOR…  1992 TAYLOR-… THE G… "RESER… Stag… 1987-2… 2       cl_31  
#> 6 19627977   PARKIN…  1993 PARKIN-… INFLA… "PRICE… Stag… 1987-2… 1       cl_48  
#> # … with 18 more rows
#> #
#> # Edge Data: 125 × 5
#>    from    to   weight Source   Target    
#>   <int> <int>    <dbl> <chr>    <chr>     
#> 1     8    20 0.000133 10729971 1111111143
#> 2     5     8 0.000253 10729971 1111111176
#> 3     8    19 0.000992 10729971 15559333  
#> # … with 122 more rows
#> 
#> $`1988-2002`
#> # A tbl_graph: 32 nodes and 231 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 32 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 25481516   HAKES-…  1988 HAKES-D… EVIDE… "REVIE… Stag… 1988-2… 1       cl_48  
#> 2 6714008    HELLIW…  1988 HELLIWE… COMPA… "JOURN… Stag… 1988-2… 3       cl_49  
#> 3 1717556    OLSON-M  1988 OLSON-M… THE P… "JOURN… Stag… 1988-2… 3       cl_49  
#> 4 214927     BALL-L   1991 BALL-L-… THE G… "JOURN… Stag… 1988-2… 1       cl_48  
#> 5 1111111176 TAYLOR…  1992 TAYLOR-… THE G… "RESER… Stag… 1988-2… 1       cl_48  
#> 6 19627977   PARKIN…  1993 PARKIN-… INFLA… "PRICE… Stag… 1988-2… 1       cl_48  
#> # … with 26 more rows
#> #
#> # Edge Data: 231 × 5
#>    from    to   weight Source    Target    
#>   <int> <int>    <dbl> <chr>     <chr>     
#> 1    26    31 0.000840 101742296 105203321 
#> 2    20    31 0.000391 101742296 1111111143
#> 3    19    31 0.00251  101742296 15559333  
#> # … with 228 more rows
#> 
#> $`1989-2003`
#> # A tbl_graph: 34 nodes and 305 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 34 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 214927     BALL-L   1991 BALL-L-… THE G… JOURNA… Stag… 1989-2… 2       cl_48  
#> 2 1111111176 TAYLOR…  1992 TAYLOR-… THE G… RESERV… Stag… 1989-2… 2       cl_48  
#> 3 19627977   PARKIN…  1993 PARKIN-… INFLA… PRICE … Stag… 1989-2… 2       cl_48  
#> 4 2207578    BALL-L   1995 BALL-L-… RELAT… THE QU… Stag… 1989-2… 2       cl_48  
#> 5 10729971   BALL-L   1995 BALL-L-… TIME-… JOURNA… Stag… 1989-2… 2       cl_48  
#> 6 24669774   ROTEMB…  1996 ROTEMBE… IMPER… JOURNA… Stag… 1989-2… 3       cl_49  
#> # … with 28 more rows
#> #
#> # Edge Data: 305 × 5
#>    from    to   weight Source    Target    
#>   <int> <int>    <dbl> <chr>     <chr>     
#> 1    24    29 0.000796 101742296 105203321 
#> 2    18    29 0.000348 101742296 1111111143
#> 3    17    29 0.00233  101742296 15559333  
#> # … with 302 more rows
#> 
#> $`1990-2004`
#> # A tbl_graph: 39 nodes and 431 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 39 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 214927     BALL-L   1991 BALL-L-… THE G… JOURNA… Stag… 1990-2… 2       cl_48  
#> 2 1111111176 TAYLOR…  1992 TAYLOR-… THE G… RESERV… Stag… 1990-2… 2       cl_48  
#> 3 19627977   PARKIN…  1993 PARKIN-… INFLA… PRICE … Stag… 1990-2… 2       cl_48  
#> 4 2207578    BALL-L   1995 BALL-L-… RELAT… THE QU… Stag… 1990-2… 2       cl_48  
#> 5 10729971   BALL-L   1995 BALL-L-… TIME-… JOURNA… Stag… 1990-2… 2       cl_48  
#> 6 24669774   ROTEMB…  1996 ROTEMBE… IMPER… JOURNA… Stag… 1990-2… 3       cl_49  
#> # … with 33 more rows
#> #
#> # Edge Data: 431 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    29    38 0.00101  101742279 101742296
#> 2    37    38 0.000143 101742279 101777013
#> 3    38    39 0.000225 101742279 106566092
#> # … with 428 more rows
#> 
#> $`1991-2005`
#> # A tbl_graph: 47 nodes and 663 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 47 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 214927     BALL-L   1991 BALL-L-… THE G… JOURNA… Stag… 1991-2… 2       cl_48  
#> 2 1111111176 TAYLOR…  1992 TAYLOR-… THE G… RESERV… Stag… 1991-2… 2       cl_48  
#> 3 19627977   PARKIN…  1993 PARKIN-… INFLA… PRICE … Stag… 1991-2… 2       cl_48  
#> 4 2207578    BALL-L   1995 BALL-L-… RELAT… THE QU… Stag… 1991-2… 2       cl_48  
#> 5 10729971   BALL-L   1995 BALL-L-… TIME-… JOURNA… Stag… 1991-2… 2       cl_48  
#> 6 24669774   ROTEMB…  1996 ROTEMBE… IMPER… JOURNA… Stag… 1991-2… 3       cl_49  
#> # … with 41 more rows
#> #
#> # Edge Data: 663 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    29    38 0.000901 101742279 101742296
#> 2    37    38 0.000125 101742279 101777013
#> 3    38    41 0.000122 101742279 106371053
#> # … with 660 more rows
#> 
#> $`1992-2006`
#> # A tbl_graph: 50 nodes and 768 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 50 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 1111111176 TAYLOR…  1992 TAYLOR-… THE G… RESERV… Stag… 1992-2… 2       cl_48  
#> 2 19627977   PARKIN…  1993 PARKIN-… INFLA… PRICE … Stag… 1992-2… 2       cl_48  
#> 3 2207578    BALL-L   1995 BALL-L-… RELAT… THE QU… Stag… 1992-2… 2       cl_48  
#> 4 10729971   BALL-L   1995 BALL-L-… TIME-… JOURNA… Stag… 1992-2… 2       cl_48  
#> 5 24669774   ROTEMB…  1996 ROTEMBE… IMPER… JOURNA… Stag… 1992-2… 3       cl_49  
#> 6 43590636   BERNAN…  1997 BERNANK… SYSTE… BROOKI… Stag… 1992-2… 3       cl_49  
#> # … with 44 more rows
#> #
#> # Edge Data: 768 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    28    37 0.000882 101742279 101742296
#> 2    36    37 0.000123 101742279 101777013
#> 3    37    40 0.000115 101742279 106371053
#> # … with 765 more rows
#> 
#> $`1993-2007`
#> # A tbl_graph: 55 nodes and 967 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 55 × 10 (active)
#>   ID_Art   Author       Year Author… Title Journal Type  time_w… cluste… intert…
#>   <chr>    <chr>       <int> <chr>   <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 19627977 PARKIN-M     1993 PARKIN… INFL… PRICE … Stag… 1993-2… 3       cl_48  
#> 2 2207578  BALL-L       1995 BALL-L… RELA… THE QU… Stag… 1993-2… 3       cl_48  
#> 3 10729971 BALL-L       1995 BALL-L… TIME… JOURNA… Stag… 1993-2… 3       cl_48  
#> 4 24669774 ROTEMBERG-J  1996 ROTEMB… IMPE… JOURNA… Stag… 1993-2… 2       cl_49  
#> 5 43590636 BERNANKE-B   1997 BERNAN… SYST… BROOKI… Stag… 1993-2… 2       cl_49  
#> 6 8456979  DELONG-B     1997 DELONG… AMER… REDUCI… Stag… 1993-2… 3       cl_48  
#> # … with 49 more rows
#> #
#> # Edge Data: 967 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    27    36 0.000857 101742279 101742296
#> 2    35    36 0.000119 101742279 101777013
#> 3    36    39 0.000110 101742279 106371053
#> # … with 964 more rows
#> 
#> $`1994-2008`
#> # A tbl_graph: 55 nodes and 985 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 55 × 10 (active)
#>   ID_Art   Author       Year Author… Title Journal Type  time_w… cluste… intert…
#>   <chr>    <chr>       <int> <chr>   <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 2207578  BALL-L       1995 BALL-L… RELA… THE QU… Stag… 1994-2… 3       cl_48  
#> 2 10729971 BALL-L       1995 BALL-L… TIME… JOURNA… Stag… 1994-2… 3       cl_48  
#> 3 24669774 ROTEMBERG-J  1996 ROTEMB… IMPE… JOURNA… Stag… 1994-2… 2       cl_49  
#> 4 43590636 BERNANKE-B   1997 BERNAN… SYST… BROOKI… Stag… 1994-2… 2       cl_49  
#> 5 8456979  DELONG-B     1997 DELONG… AMER… REDUCI… Stag… 1994-2… 3       cl_48  
#> 6 27334211 TAYLOR-J     1997 TAYLOR… AMER… REDUCI… Stag… 1994-2… 3       cl_48  
#> # … with 49 more rows
#> #
#> # Edge Data: 985 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    26    35 0.000848 101742279 101742296
#> 2    34    35 0.000116 101742279 101777013
#> 3    35    38 0.000110 101742279 106371053
#> # … with 982 more rows
#> 
#> $`1995-2009`
#> # A tbl_graph: 60 nodes and 1151 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 60 × 10 (active)
#>   ID_Art   Author       Year Author… Title Journal Type  time_w… cluste… intert…
#>   <chr>    <chr>       <int> <chr>   <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 2207578  BALL-L       1995 BALL-L… RELA… THE QU… Stag… 1995-2… 3       cl_48  
#> 2 10729971 BALL-L       1995 BALL-L… TIME… JOURNA… Stag… 1995-2… 3       cl_48  
#> 3 24669774 ROTEMBERG-J  1996 ROTEMB… IMPE… JOURNA… Stag… 1995-2… 2       cl_49  
#> 4 43590636 BERNANKE-B   1997 BERNAN… SYST… BROOKI… Stag… 1995-2… 2       cl_49  
#> 5 8456979  DELONG-B     1997 DELONG… AMER… REDUCI… Stag… 1995-2… 3       cl_48  
#> 6 27334211 TAYLOR-J     1997 TAYLOR… AMER… REDUCI… Stag… 1995-2… 3       cl_48  
#> # … with 54 more rows
#> #
#> # Edge Data: 1,151 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    26    35 0.000838 101742279 101742296
#> 2    34    35 0.000115 101742279 101777013
#> 3    35    38 0.000108 101742279 106371053
#> # … with 1,148 more rows
#> 
#> $`1996-2010`
#> # A tbl_graph: 60 nodes and 1163 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 60 × 10 (active)
#>   ID_Art   Author       Year Author… Title Journal Type  time_w… cluste… intert…
#>   <chr>    <chr>       <int> <chr>   <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 24669774 ROTEMBERG-J  1996 ROTEMB… IMPE… JOURNA… Stag… 1996-2… 2       cl_49  
#> 2 43590636 BERNANKE-B   1997 BERNAN… SYST… BROOKI… Stag… 1996-2… 2       cl_49  
#> 3 8456979  DELONG-B     1997 DELONG… AMER… REDUCI… Stag… 1996-2… 3       cl_48  
#> 4 27334211 TAYLOR-J     1997 TAYLOR… AMER… REDUCI… Stag… 1996-2… 3       cl_48  
#> 5 60300900 CHARI-V      1998 CHARI-… EXPE… JOURNA… Stag… 1996-2… 3       cl_48  
#> 6 94940407 HETZEL-R     1998 HETZEL… ARTH… FEDERA… Stag… 1996-2… 1       cl_31  
#> # … with 54 more rows
#> #
#> # Edge Data: 1,163 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    24    33 0.000838 101742279 101742296
#> 2    32    33 0.000115 101742279 101777013
#> 3    33    36 0.000108 101742279 106371053
#> # … with 1,160 more rows
#> 
#> $`1997-2011`
#> # A tbl_graph: 60 nodes and 1183 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 60 × 10 (active)
#>   ID_Art   Author      Year Author_… Title Journal Type  time_w… cluste… intert…
#>   <chr>    <chr>      <int> <chr>    <chr> <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 43590636 BERNANKE-B  1997 BERNANK… SYST… "BROOK… Stag… 1997-2… 2       cl_49  
#> 2 8456979  DELONG-B    1997 DELONG-… AMER… "REDUC… Stag… 1997-2… 3       cl_48  
#> 3 27334211 TAYLOR-J    1997 TAYLOR-… AMER… "REDUC… Stag… 1997-2… 3       cl_48  
#> 4 60300900 CHARI-V     1998 CHARI-V… EXPE… "JOURN… Stag… 1997-2… 3       cl_48  
#> 5 94940407 HETZEL-R    1998 HETZEL-… ARTH… "FEDER… Stag… 1997-2… 1       cl_31  
#> 6 17355017 MAYER-T     1998 MAYER-T… MONE… ""      Stag… 1997-2… 3       cl_48  
#> # … with 54 more rows
#> #
#> # Edge Data: 1,183 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    23    32 0.000836 101742279 101742296
#> 2    31    32 0.000114 101742279 101777013
#> 3    32    35 0.000107 101742279 106371053
#> # … with 1,180 more rows
#> 
#> $`1998-2012`
#> # A tbl_graph: 59 nodes and 1199 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 59 × 10 (active)
#>   ID_Art   Author   Year Author_… Title   Journal  Type  time_w… cluste… intert…
#>   <chr>    <chr>   <int> <chr>    <chr>   <chr>    <chr> <chr>   <chr>   <chr>  
#> 1 60300900 CHARI-V  1998 CHARI-V… EXPECT… "JOURNA… Stag… 1998-2… 1       cl_31  
#> 2 94940407 HETZEL…  1998 HETZEL-… ARTHUR… "FEDERA… Stag… 1998-2… 2       cl_82  
#> 3 17355017 MAYER-T  1998 MAYER-T… MONETA… ""       Stag… 1998-2… 1       cl_31  
#> 4 73951174 IRELAN…  1999 IRELAND… DOES T… "JOURNA… Stag… 1998-2… 1       cl_31  
#> 5 24247525 TAYLOR…  1999 TAYLOR-… A HIST… "MONETA… Stag… 1998-2… 3       cl_49  
#> 6 95020640 CHRIST…  2000 CHRISTI… THE EX… ""       Stag… 1998-2… 2       cl_82  
#> # … with 53 more rows
#> #
#> # Edge Data: 1,199 × 5
#>    from    to   weight Source    Target   
#>   <int> <int>    <dbl> <chr>     <chr>    
#> 1    20    29 0.000835 101742279 101742296
#> 2    28    29 0.000112 101742279 101777013
#> 3    29    32 0.000105 101742279 106371053
#> # … with 1,196 more rows
#> 
#> $`1999-2013`
#> # A tbl_graph: 64 nodes and 1515 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 64 × 10 (active)
#>   ID_Art     Author   Year Author_… Title  Journal Type  time_w… cluste… intert…
#>   <chr>      <chr>   <int> <chr>    <chr>  <chr>   <chr> <chr>   <chr>   <chr>  
#> 1 73951174   IRELAN…  1999 IRELAND… DOES … "JOURN… Stag… 1999-2… 1       cl_31  
#> 2 24247525   TAYLOR…  1999 TAYLOR-… A HIS… "MONET… Stag… 1999-2… 2       cl_49  
#> 3 95020640   CHRIST…  2000 CHRISTI… THE E… ""      Stag… 1999-2… 3       cl_86  
#> 4 76064614   CLARID…  2000 CLARIDA… MONET… "THE Q… Stag… 1999-2… 2       cl_49  
#> 5 15559333   LANSIN…  2000 LANSING… EXPLO… "RESEA… Stag… 1999-2… 1       cl_31  
#> 6 1111111143 LANSIN…  2000 LANSING… LEARN… ""      Stag… 1999-2… 1       cl_31  
#> # … with 58 more rows
#> #
#> # Edge Data: 1,515 × 5
#>    from    to    weight Source    Target   
#>   <int> <int>     <dbl> <chr>     <chr>    
#> 1    17    26 0.000785  101742279 101742296
#> 2    25    26 0.000105  101742279 101777013
#> 3    26    29 0.0000994 101742279 106371053
#> # … with 1,512 more rows
#>